% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_otu}
\alias{add_otu}
\title{Add OTUs as row names}
\usage{
add_otu(df, add_column = FALSE)
}
\arguments{
\item{df}{The input MetaPhlAn data frame.}

\item{add_column}{\code{Logical} value indicating whether to add a OTU column
to the data frame.}
}
\value{
The data frame with OTU row names added.
}
\description{
This function adds numbered Operational Taxonomic Units (OTUs)
to the row names of a data frame. Optionally can also add a OTU column to
the data frame.
}
\examples{
df <- data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
add_otu(df) # Adds OTU row names to the data frame
add_otu(df, add_column = TRUE) # Adds OTU column and row names to the data frame

}
\author{
Jérémy Rotzetter
}
