% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Remove pattern from column names}
\usage{
clean_colnames(df, pattern)
}
\arguments{
\item{df}{A data frame or matrix.}

\item{pattern}{A \code{character} string with the pattern to match.}
}
\value{
Returns the input table with the specified pattern removed from the
column names.
}
\description{
Function to remove a pattern from column names.
}
\examples{
colnames(merged_abundance_profiles)

names_cleaned <- clean_colnames(merged_abundance_profiles, "SRS0144\\\\d{2}.")

colnames(names_cleaned)

}
\author{
Jérémy Rotzetter
}
