% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Only keep shared samples between metadata and MetaPhlAn profile}
\usage{
clean_metadata(metadata, mtphlan_profile, sample_column, remove_spaces = TRUE)
}
\arguments{
\item{metadata}{A data frame containing metadata information.}

\item{mtphlan_profile}{A data frame containing metaphlan profile information.
Samples are columns.}

\item{sample_column}{A \code{character} string. The column in the metadata
containing the sample names. Should match column names of the MetaPhlAn
profile.}

\item{remove_spaces}{Should spaces (" ") be replaced with "_" in the column
names?}
}
\value{
The cleaned metadata as a data frame.
}
\description{
The \code{clean_metadata()} function is designed to clean and prepare
metadata and a MetaPhlAn profile for further analysis by only keeping shared
samples.
}
\examples{
metadata <- data.frame(
  Sample_name = c("A", "B", "C", "D", "F", "G"),
  Sex = c(rep(c("F", "M")))
)

mtphlan_profile <- data.frame(
  clade_name = c("x", "y", "z"),
  A = c(1, 2, 3),
  B = c(4, 5, 6),
  D = c(7, 8, 9),
  Z = c(10, 11, 12)
)

cleaned_metadata <- clean_metadata(metadata, mtphlan_profile, "Sample_name")

print(cleaned_metadata)

}
\author{
Jérémy Rotzetter
}
