% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_to_rownames}
\alias{col_to_rownames}
\title{Set a column as row names}
\usage{
col_to_rownames(df, column)
}
\arguments{
\item{df}{The input data frame.}

\item{column}{A \code{character} string. The name of the column to be set as
row names.}
}
\value{
The input data frame with the specified column set as row names.
}
\description{
This function sets a specified column as the row names of a
data frame.
}
\examples{
df <- data.frame(
  clade_name = c("A", "B", "C"),
  col1 = c(10, 20, 5),
  col2 = c(8, 15, 3),
  col3 = c(64, 89, 32)
)
col_to_rownames(df, "clade_name")

}
