% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_threshold}
\alias{filter_threshold}
\title{Apply threshold filtering to a data frame}
\usage{
filter_threshold(df, threshold, merged_profiles = TRUE)
}
\arguments{
\item{df}{The input data frame of merged MetaPhlAn profiles.}

\item{threshold}{The threshold value for filtering.}

\item{merged_profiles}{\code{Logical}; if \code{TRUE} (the default) the file to be
loaded is assumed to be multiple merged MetaPhlAn profiles.}
}
\value{
The filtered data frame.
}
\description{
This function applies threshold filtering to a data frame,
summing values below the threshold for each column, storing the sum in the
new 'Other' row, but only if all values within a row (taxon) across all
columns are below the threshold.
}
\details{
Function can be applied to a single MetaPhlAn profile by setting
\code{merged_profiles = FALSE}. Likewise also works with merged profiles
by slicing the data frame to only the clade_name and the numeric sample
column of interest. In both cases a \code{clade_name} column is required.
}
\examples{
df <- data.frame(
  clade_name = c("A", "B", "C", "D"),
  col1 = c(10, 20, 5, 65),
  col2 = c(8, 15, 3, 38),
  col3 = c(5, 35, 4, 6)
)
print(df)
df_thresh <- filter_threshold(df, 11)
print(df_thresh)

}
\author{
Jérémy Rotzetter
}
