% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize_str}
\alias{capitalize_str}
\alias{lowercase_str}
\title{Modify the case of the first letter in each word of a character string}
\usage{
capitalize_str(character_string)

lowercase_str(character_string)
}
\arguments{
\item{character_string}{A \code{character} vector containing the strings to be
modified.}
}
\value{
A character vector with the first letter of each word modified to
either uppercase or lowercase.
}
\description{
These functions take a character string as input and return a
new character string with the first letter of each word modified to either
uppercase or lowercase.
}
\examples{
capitalize_str("hello world") # Returns "Hello World"
capitalize_str(c("hello", "world")) # Returns c("Hello", "World")
lowercase_str("Hello World") # Returns "hello world"
lowercase_str(c("Hello", "World")) # Returns c("hello", "world")

}
