% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_taxa_table}
\alias{get_taxa_table}
\title{Get taxonomic table}
\usage{
get_taxa_table(
  df,
  taxa_lvl = NULL,
  taxa_are_rows = FALSE,
  use_taxa_names = FALSE
)
}
\arguments{
\item{df}{The input data frame.}

\item{taxa_lvl}{The taxonomic level to extract ('kingdom', 'phylum', 'class',
'order', 'family', 'genus', 'species' or 't' (SGB)). First letter abbreviations
(e.g., 's') are also accepted. Is deactivated with \code{NULL} by default.}

\item{taxa_are_rows}{\code{Logical}; if \code{TRUE}, taxa information is in the row
names of the data frame, otherwise it is assumed to be in the 'clade_name'
column.}

\item{use_taxa_names}{\code{Logical}; if \code{TRUE}, assign taxa names as row names,
otherwise assign Otu as row names.}
}
\value{
A taxonomic matrix.
}
\description{
Function to extract taxonomic information from a data frame and
prepare it for phyloseq.
}
\note{
Will work with unfiltered as well as filtered to a specific taxonomic
level input data frames.
}
\examples{
taxmat <- get_taxa_table(species_only, taxa_lvl = "s")
head(taxmat)

}
\author{
Jérémy Rotzetter
}
