% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{list_example}
\alias{list_example}
\title{List example files}
\source{
This function was adapted from \code{\link[readxl:readxl_example]{readxl::readxl_example()}} and
originally created by \emph{Hadley Wickham & Jennifer Bryan}.
}
\usage{
list_example(path = NULL)
}
\arguments{
\item{path}{The name of an example file. If not provided, the function lists
the available example files.}
}
\value{
A character vector.
}
\description{
This function lists available example files, found in the \code{inst/extdata}
directory.
}
\examples{
list_example() # Lists example files contained in the package
list_example("merged_abundance_table.txt") # Lists the path of the specified file

}
