% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_metaphlan_profile}
\alias{load_metaphlan_profile}
\title{Load MetaPhlAn profile(s)}
\usage{
load_metaphlan_profile(path, merged_profiles = TRUE)
}
\arguments{
\item{path}{A \code{character} string containing the path to the file to read.}

\item{merged_profiles}{\code{Logical}; if \code{TRUE} (the default) the file to be
loaded is assumed to be multiple merged MetaPhlAn profiles.}
}
\value{
A data frame containing the MetaPhlAn results.
}
\description{
Function to load a MetaPhlAn profile from a file path.
}
\examples{
path <- list_example("merged_abundance_table.txt")
data <- load_metaphlan_profile(path = path)
head(data)

}
\author{
Jérémy Rotzetter
}
