% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{merged_abundance_profiles}
\alias{merged_abundance_profiles}
\title{Merged MetaPhlAn profiles data}
\format{
\subsection{\code{merged_abundance_profiles}}{

A data frame with 59 rows and 7 columns:
\describe{
\item{clade_name}{The taxonomic lineage of the taxon reported on this line.
Taxon names are prefixed with one-letter codes to help indicate their rank.}
\item{SRS0144...}{The sample names for which sample-specific abundance
profiles are found in a column.}
}
}
}
\source{
\url{https://github.com/biobakery/biobakery/wiki/MetaPhlAn-4.1}
}
\usage{
merged_abundance_profiles
}
\description{
Example data of multiple MetaPhlAn profiles merged into one file.
}
\keyword{datasets}
