% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan_to_phyloseq.R
\name{metaphlan_to_phyloseq}
\alias{metaphlan_to_phyloseq}
\title{Convert MetaPhlAn profile to phyloseq object}
\usage{
metaphlan_to_phyloseq(
  mtphlan_profile,
  taxa_lvl,
  metadata = NULL,
  sample_column = NULL,
  use_taxa_names = FALSE,
  merged_profiles = TRUE
)
}
\arguments{
\item{mtphlan_profile}{The MetaPhlAn profile to be converted. It can be
either a file path or a data frame of MetaPhlAn profile(s).}

\item{taxa_lvl}{The taxonomic level to filter the profile to. Valid options
are 'kingdom', 'phylum', 'class', 'order', 'family', 'genus', or 'species'.
First letter abbreviations are also accepted.}

\item{metadata}{Optional metadata for the samples. If provided, it should be
a data frame.}

\item{sample_column}{\code{character} string. The column in the metadata
containing the sample names. Should match column names of the MetaPhlAn
profile.}

\item{use_taxa_names}{\code{Logical} indicating whether to use taxonomic names
instead of OTUs in the resulting phyloseq object. Default is \code{FALSE}.}

\item{merged_profiles}{\code{Logical}; if \code{TRUE} (the default) the file to be
loaded is assumed to be multiple merged MetaPhlAn profiles.}
}
\value{
A phyloseq object representing the MetaPhlAn profile.
}
\description{
This function converts a MetaPhlAn profile to a phyloseq object,
which is a common data structure used in microbiome analysis.
}
\note{
If \code{mtphlan_profile} is an already loaded object, it should be filtered to
the chosen taxonomic rank prior to calling \code{metaphlan_to_phyloseq()}. The
clade names should also not be shortened, otherwise the taxonomic table
cannot be created. Should you wish to use the taxonomic names of the
specified rank as the row names in the phyloseq object, set
\code{use_taxa_names = TRUE}.
}
\examples{
metaphlan_to_phyloseq(species_only, taxa_lvl = "s")

}
\author{
Jérémy Rotzetter
}
