% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlanToPhyloseq-package.R
\docType{package}
\name{metaphlanToPhyloseq-package}
\alias{metaphlanToPhyloseq}
\alias{metaphlanToPhyloseq-package}
\title{metaphlanToPhyloseq: Convert MetaPhlAn taxonomic profiles to phyloseq objects}
\description{
Simple package to load MetaPhlAn taxonomic profiles, clean or filter them based on specified taxonomic rank or abundance threshold, and create a phyloseq object from them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrotzetter/metaphlanToPhyloseq}
  \item Report bugs at \url{https://github.com/jrotzetter/metaphlanToPhyloseq/issues}
}

}
\author{
\strong{Maintainer}: Jérémy Rotzetter \email{first.last@example.com} (\href{https://orcid.org/0009-0007-6076-7077}{ORCID})

}
\keyword{internal}
