% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{single_abundance_profile}
\alias{single_abundance_profile}
\title{SRS014470-Tongue_dorsum_profile data}
\format{
\subsection{\code{single_abundance_profile}}{

A data frame with 15 rows and 4 columns:
\describe{
\item{clade_name}{The taxonomic lineage of the taxon reported on this line.
Taxon names are prefixed with one-letter codes to help indicate their rank.}
\item{NCBI_tax_id}{The NCBI-equivalent taxon IDs of the named taxa from
clade_name.}
\item{relative_abundance}{The taxon's relative abundance in \%.}
\item{additional_species}{Additional species names for cases where the
metagenome profile contains clades that represent multiple species. The
species listed in column 1 is the representative species in such cases.}
}
}
}
\source{
\url{https://github.com/biobakery/biobakery/wiki/MetaPhlAn-4.1}
}
\usage{
single_abundance_profile
}
\description{
Example data of a single MetaPhlAn profile output file, containing the
computed taxon abundances.
}
\keyword{datasets}
