% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{species_only}
\alias{species_only}
\title{Merged MetaPhlAn profiles data to species level}
\format{
\subsection{\code{species_only}}{

A data frame with 12 rows and 7 columns:
\describe{
\item{clade_name}{The taxonomic lineage of the taxon reported on this line.
Taxon names are prefixed with one-letter codes to help indicate their rank.}
\item{SRS0144...}{The sample names for which sample-specific abundance
profiles are found in a column.}
}
}
}
\source{
\url{https://github.com/biobakery/biobakery/wiki/MetaPhlAn-4.1}
}
\usage{
species_only
}
\description{
Example data of multiple MetaPhlAn profiles which were merged into one file,
and then filtered with \code{filter_taxa_lvl(df, taxa_lvl = "s")} to only include
species-level.
}
\keyword{datasets}
