% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{shorten_clade_names}
\alias{shorten_clade_names}
\title{Shorten clade names in a dataset to chosen taxonomic level}
\usage{
shorten_clade_names(
  data,
  taxa_lvl,
  apply_to_colnames = TRUE,
  selected_cols = NULL
)
}
\arguments{
\item{data}{The input dataset. Assumes clade names are of the structure
\emph{\strong{k__Bacteria|p__Firmicutes|c__Bacilli|o__Lactobacillales|f__Lactobacillaceae}}
and found in the 'clade_name' column.}

\item{taxa_lvl}{The taxonomic level at which the clade names should be
shortened. Valid options include 'kingdom', 'phylum', 'class', 'order',
'family', 'genus', or 'species'. First letter abbreviations (e.g., 's')
are also accepted.}

\item{apply_to_colnames}{\code{Logical} indicating whether the shortening should
be applied to column names or row values. Default is TRUE.}

\item{selected_cols}{A \code{character} vector specifying the columns to which
the shortening should be applied. If \code{NULL}, the shortening is applied to all
columns. Default is \code{NULL}.}
}
\value{
The dataset with the clade names shortened based on the specified
taxonomic level.
}
\description{
This function shortens the taxonomic names of clades in a given
dataset based on a specified taxonomic level, using '|' as a separator.
}
\details{
This function uses the dplyr package for data manipulation.
}
\examples{
head(species_only$clade_name)

taxa_shortened <- shorten_clade_names(
  species_only,
  "species",
  apply_to_colnames = FALSE,
  selected_cols = "clade_name")

  head(taxa_shortened$clade_name)

}
\author{
Jérémy Rotzetter
}
